refactor: rename workflows.ts to workflowService.ts
- Renamed packages/lib/server/service/workflows.ts to workflowService.ts
- Updated import statements in teamService.ts and related test files
- Updated vi.mock statements in test files to use new file path
